# backend/main-fiber/docs/swagger.yaml
basePath: /api
definitions:
...
  api.abilitiesResp:
    properties:
      abilities:
        additionalProperties:
          type: boolean
        type: object
      org_path:
        type: string
      version:
        type: string
    type: object
...
  api.loginReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  api.loginResp:
    properties:
      access_token:
        type: string
    type: object
...
info:
  contact: {}
  description: This is my API
  title: My API
  version: "1.0"
paths:
  /api/abilities:
    get:
      consumes:
      - application/json
      description: Returns allowed actions for the user in the specified org path
      parameters:
      - description: Organization path
        in: query
        name: org_path
        required: true
        type: string
      - description: Comma-separated actions to check
        in: query
        name: actions
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.abilitiesResp'
      summary: Get user abilities for an org path
      tags:
      - abilities
  /api/abilities/where:
    get:
      consumes:
      - application/json
      description: Returns org paths where the user has permission for the specified
        action
      parameters:
      - description: Action to check
        in: query
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List org paths where user can perform an action
      tags:
      - abilities
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns JWT access token
      parameters:
      - description: Login credentials
        in: body
        name: loginReq
        required: true
        schema:
          $ref: '#/definitions/api.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.loginResp'
      summary: Login
      tags:
      - auth
  /api/auth/me:
    get:
      consumes:
      - application/json
      description: Returns claims of the authenticated user
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get current user info
      tags:
      - auth
...
swagger: "2.0"